+++ src/grandorgue/GODocument.h
@@ -11,15 +11,15 @@
 #include <wx/string.h>
 
 #include "document-base/GODocumentBase.h"
-#include "midi/GOMidiCallback.h"
 #include "midi/GOMidiListener.h"
 #include "midi/dialog-creator/GOMidiDialogCreator.h"
+#include "midi/events/GOMidiCallback.h"
 #include "threading/GOMutex.h"
 
 class GOOrganController;
 class GOMidiShortcutReceiver;
 class GOMidiEvent;
-class GOMidiReceiverBase;
+class GOMidiReceiver;
 class GOMidiSender;
 class GOOrgan;
 class GOProgressDialog;
@@ -69,7 +69,7 @@ public:
     void *element,
     const wxString &title,
     const wxString &dialogSelector,
-    GOMidiReceiverBase *event,
+    GOMidiReceiver *event,
     GOMidiSender *sender,
     GOMidiShortcutReceiver *key,
     GOMidiSender *division = nullptr,