+++ src/grandorgue/gui/dialogs/midi-event/GOMidiEventRecvTab.cpp
@@ -15,10 +15,11 @@
 #include <wx/stattext.h>
 #include <wx/tglbtn.h>
 
-#include "GOOrganController.h"
-#include "GORodgers.h"
 #include "config/GOConfig.h"
-#include "midi/GOMidiEvent.h"
+#include "midi/events/GOMidiEvent.h"
+#include "midi/events/GORodgers.h"
+
+#include "GOOrganController.h"
 
 BEGIN_EVENT_TABLE(GOMidiEventRecvTab, wxPanel)
 EVT_TOGGLEBUTTON(ID_LISTEN_SIMPLE, GOMidiEventRecvTab::OnListenSimpleClick)
@@ -31,7 +32,7 @@ EVT_CHOICE(ID_EVENT, GOMidiEventRecvTab::OnTypeChange)
 END_EVENT_TABLE()
 
 GOMidiEventRecvTab::GOMidiEventRecvTab(
-  wxWindow *parent, GOMidiReceiverBase *event, GOConfig &config)
+  wxWindow *parent, GOMidiReceiver *event, GOConfig &config)
   : wxPanel(parent, wxID_ANY),
     m_MidiIn(config.m_MidiIn),
     m_MidiMap(config.GetMidiMap()),