+++ src/grandorgue/midi/elements/GOMidiReceiver.h
@@ -5,21 +5,25 @@
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
 
-#ifndef GOMIDIRECEIVERBASE_H
-#define GOMIDIRECEIVERBASE_H
+#ifndef GOMIDIRECEIVER_H
+#define GOMIDIRECEIVER_H
 
 #include <cstdint>
 
-#include "GOMidiMatchType.h"
-#include "GOMidiReceiverEventPatternList.h"
+#include "midi/events/GOMidiMatchType.h"
+#include "midi/events/GOMidiReceiverEventPatternList.h"
+
+#include "GOMidiElement.h"
 #include "GOTime.h"
 
+class GOConfig;
 class GOConfigReader;
 class GOConfigWriter;
 class GOMidiEvent;
 class GOMidiMap;
 
-class GOMidiReceiverBase : public GOMidiReceiverEventPatternList {
+class GOMidiReceiver : public GOMidiReceiverEventPatternList,
+                       public GOMidiElement {
 public:
   constexpr static unsigned KEY_MAP_SIZE = 128;
   using KeyMap = uint8_t[KEY_MAP_SIZE];
@@ -31,6 +35,8 @@ private:
     int key;
   } midi_internal_match;
 
+  const GOConfig &r_config;
+
   int m_ElementID;
   std::vector<GOTime> m_last;
   std::vector<midi_internal_match> m_Internal;
@@ -41,16 +47,21 @@ private:
   unsigned createInternal(unsigned device);
 
 protected:
-  virtual int GetTranspose() const { return 0; }
+  int GetTranspose() const;
 
 public:
-  GOMidiReceiverBase(GOMidiReceiverType type);
+  GOMidiReceiver(GOConfig &config, GOMidiReceiverType type);
 
-  virtual void Load(GOConfigReader &cfg, const wxString &group, GOMidiMap &map);
+  void Load(GOConfigReader &cfg, const wxString &group, GOMidiMap &map);
   void Save(GOConfigWriter &cfg, const wxString &group, GOMidiMap &map) const;
+
+  void ToYaml(YAML::Node &yamlNode, GOMidiMap &map) const override;
+  void FromYaml(const YAML::Node &yamlNode, GOMidiMap &map) override;
+
+public:
   void PreparePlayback();
 
-  void SetElementID(int id);
+  void SetElementID(int id) { m_ElementID = id; }
 
   GOMidiMatchType Match(const GOMidiEvent &e);
   GOMidiMatchType Match(const GOMidiEvent &e, int &value);